dalil_group/requirements.txt
# LinguaEval Platform - Python Dependencies

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9

# HTTP Client
httpx>=0.27.0
aiofiles>=24.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Caching & Async
redis>=5.0.0
celery>=5.3.0

# Vector Database & Embeddings (Knowledge Agent)
chromadb>=1.0.0
sentence-transformers>=3.0.0

# Document Processing & Reports
python-docx>=1.1.0
fpdf2>=2.7.0
python-pptx>=1.0.0
PyPDF2>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# LLM Providers (optional)
openai>=1.0.0
anthropic>=0.30.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.27.0

# CI/CD & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Production Database
psycopg2-binary>=2.9.0  # PostgreSQL
pg8000>=1.29.0  # Alternative PostgreSQL driver

# Web Scraping for RAG
requests>=2.31.0
beautifulsoup4>=4.12.0

# Fine-tuning & ML
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0  # LoRA/QLoRA
bitsandbytes>=0.40.0  # Quantization
accelerate>=0.20.0  # Distributed training
datasets>=2.14.0  # Dataset handling

# Production Server
gunicorn>=21.0.0
python-dotenv>=1.0.0

# Monitoring & Logging
prometheus-client>=0.17.0
structlog>=23.1.0

# CLI
click>=8.1.0

# Utilities
python-dotenv>=1.0.0
python-json-logger>=2.0.0

